Initial stab at adding sources to the project
This commit adds some "real" sources to the subproject `:a` only. Those sources are "real" sources in the sense
that it contains sources coming from the WordPress app, but stripped down to allow compilation and processing of
resources.

We need to make it clear that the objective of this project is *not* to create a valid Android app. It's pointless
to try to execute this app as it will *never* work. The only objective is to simulate what would happen with a
large Android application.