Integrate dummy sources into all Android subprojects
This commit adds dummy sources, the same as the ones from the `:a` subproject, to all subprojects
which are Android subprojects. This means that there are still subprojects without any source.

However, this change is pretty huge. It dramatically increases build time, and also requires 12G of
RAM to execute. It's also worth noting that 12G RAM is not enough to run with `--parallel`, making
it effectively impossible to build this project using reasonably sized hardware (laptop) with parallel
execution enabled.

The script that was used to copy the sources can be found in `copy-sources.groovy`.